requirements.txt
openai
qdrant-client
pytesseract
Pillow
PyMuPDF
langchain
tiktoken
pymongo
boto3
requests
fastapi
uvicorn